refactor: move map to artist response from service to controller

+++ backend/src/main/java/com/xdpsx/music/service/impl/ArtistServiceImpl.java
@@ -2,8 +2,6 @@ package com.xdpsx.music.service.impl;
 
 import com.xdpsx.music.dto.request.params.ArtistParams;
 import com.xdpsx.music.dto.request.ArtistRequest;
-import com.xdpsx.music.dto.response.ArtistResponse;
-import com.xdpsx.music.dto.common.PageResponse;
 import com.xdpsx.music.model.entity.Artist;
 import com.xdpsx.music.exception.ResourceNotFoundException;
 import com.xdpsx.music.mapper.ArtistMapper;
@@ -17,9 +15,6 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
-import java.util.List;
-import java.util.stream.Collectors;
-
 import static com.xdpsx.music.constant.FileConstants.ARTISTS_IMG_FOLDER;
 
 @Service
@@ -30,39 +25,36 @@ public class ArtistServiceImpl implements ArtistService {
     private final ArtistRepository artistRepository;
 
     @Override
-    public PageResponse<ArtistResponse> getAllArtists(ArtistParams params) {
+    public Page<Artist> getAllArtists(ArtistParams params) {
         Pageable pageable = PageRequest.of(params.getPageNum() - 1, params.getPageSize());
-        Page<Artist> artistPage = artistRepository.findWithFilters(
+        return artistRepository.findWithFilters(
                 pageable, params.getSearch(), params.getSort(), params.getGender()
         );
-        List<ArtistResponse> responses = artistPage.getContent().stream()
-                .map(artistMapper::fromEntityToResponse)
-                .collect(Collectors.toList());
-        return PageResponse.<ArtistResponse>builder()
-                .items(responses)
-                .pageNum(artistPage.getNumber() + 1)
-                .pageSize(artistPage.getSize())
-                .totalItems(artistPage.getTotalElements())
-                .totalPages(artistPage  .getTotalPages())
-                .build();
     }
 
     @Override
-    public ArtistResponse createArtist(ArtistRequest request, MultipartFile image) {
+    public Artist getArtistById(Long id) {
+        return artistRepository.findById(id)
+                .orElse(null);
+    }
+
+    @Override
+    public Artist createArtist(ArtistRequest request, MultipartFile image) {
         Artist artist = artistMapper.fromRequestToEntity(request);
         if (image != null){
             String imageUrl = fileService.uploadFile(image, ARTISTS_IMG_FOLDER);
             artist.setAvatar(imageUrl);
         }
 
-        Artist savedArtist = artistRepository.save(artist);
-        return artistMapper.fromEntityToResponse(savedArtist);
+        return artistRepository.save(artist);
     }
 
     @Override
-    public ArtistResponse updateArtist(Long id, ArtistRequest request, MultipartFile image) {
-        Artist artist = artistRepository.findById(id)
-                .orElseThrow(() -> new ResourceNotFoundException(String.format("Not found artist with ID=%s", id)));
+    public Artist updateArtist(Long id, ArtistRequest request, MultipartFile image) {
+        Artist artist = getArtistById(id);
+        if (artist == null){
+            throw new ResourceNotFoundException(String.format("Not found artist with ID=%s", id));
+        }
         artist.setName(request.getName());
         artist.setGender(request.getGender());
         artist.setDescription(request.getDescription());
@@ -79,21 +71,16 @@ public class ArtistServiceImpl implements ArtistService {
         if (oldImage != null){
             fileService.deleteFileByUrl(oldImage);
         }
-        return artistMapper.fromEntityToResponse(savedArtist);
-    }
-
-    @Override
-    public ArtistResponse getArtistById(Long id) {
-        Artist artist = artistRepository.findById(id)
-                .orElseThrow(() -> new ResourceNotFoundException(String.format("Not found artist with ID=%s", id)));
-        return artistMapper.fromEntityToResponse(artist);
+        return savedArtist;
     }
 
     @Override
     public void deleteArtist(Long id) {
-        Artist artist = artistRepository.findById(id)
-                .orElseThrow(() -> new ResourceNotFoundException(String.format("Not found artist with ID=%s", id)));
-        fileService.deleteFileByUrl(artist.getAvatar());
+        Artist artist = getArtistById(id);
+        if (artist == null){
+            throw new ResourceNotFoundException(String.format("Not found artist with ID=%s", id));
+        }
         artistRepository.delete(artist);
+        fileService.deleteFileByUrl(artist.getAvatar());
     }
 }
